# portal/backend/openapi/bff.yaml
openapi: 3.0.3
info:
  title: OpenFGC Portal BFF API
  version: 0.2.0
  description: >-
    Stateless portal BFF endpoints, including health checks, portal-facing
    user routes, and explicit passthrough routes to consent-server. For proxied
    endpoints in this spec, response payload shape is inherited from the
    consent-management API.
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/consents:
    get:
      tags: [Passthrough]
      summary: Passthrough search consents
      description: >-
        Response payload shape is inherited from consent-management API.
      responses:
        '200':
          description: Upstream response
          content:
            application/json:
              schema:
                type: object
        '404':
          $ref: '#/components/responses/ProxyError'
        '405':
          $ref: '#/components/responses/ProxyError'
        '502':
          $ref: '#/components/responses/ProxyError'
        '503':
          $ref: '#/components/responses/ProxyError'
    post:
      tags: [Passthrough]
      summary: Passthrough create consent
      description: >-
        Response payload shape is inherited from consent-management API.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Upstream response
          content:
            application/json:
              schema:
                type: object
        '404':
          $ref: '#/components/responses/ProxyError'
        '405':
          $ref: '#/components/responses/ProxyError'
        '413':
          $ref: '#/components/responses/ProxyError'
        '502':
          $ref: '#/components/responses/ProxyError'
        '503':
          $ref: '#/components/responses/ProxyError'
components:
  schemas:
...
    ProxyError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: UPSTREAM_UNAVAILABLE
        message:
          type: string
          example: upstream unavailable
  responses:
    ProxyError:
      description: Proxy-layer error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProxyError'